\name{auction}

\alias{auction-package}

\docType{package}

\title{Auction Algorithms with R}

\description{
The auction package is a collection solvers for real-valued transportation problems.
}

\details{
\tabular{ll}{
Package: \tab auction\cr
Type: \tab Package\cr
Version: \tab 1.2.0\cr
Date: \tab 2018-02-02\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab Yes\cr
Depends: \tab Rcpp, RcppArmadillo \cr }
}
\section{How to cite this package}{
 \preformatted{@Manual{OHara2018,
 author       = {Keith O'Hara},
 title        = {{auction}: Auction Algorithms with R},
 year         = {2018},
 note         = {R package version 1.1.0.}}}
\emph{}
}

\section{License}{
auction is licensed under the GNU General Public License (GPL) version 2, or (at your option) any later version.
}

\author{Keith O'Hara}